fix: fix the check for GU:03
The previous check for the GU:03 error, user not found, was incorrect. It was using readErrors() to build an error string for the Exception, and then testing that string for a value that it would not contain.

I modified this somewhat.  I added a private helper method called isErrorType() which inspects the <Errors/> set and looks for a specific ErrorID. If it contains an error with that ID it returns true.  I then used this in place of the string_contains() test to look for the error and return null.