feat: extend SmarterUException with request/response
Troubleshooting issues with SmarterU is made several times more complicated than it needs to be because we don't capture useful information in the exception and log it anywhere.  I am to fix this by adding request and response to the exception details.  I don't want sensitive credentials going in our logs so I made sure to sanitize the request.